Calculate personal finance

+++ python_introduction/finance_calculator.py
@@ -6,7 +6,7 @@ mo_saving = income - mo_expenses
 
 annual_interest = 0.05
 
-projected_saving = (mo_saving *12) + (mo_saving * 12 * 0.05)
+projected_saving = int((mo_saving *12) + (mo_saving * 12 * 0.05))
 
 print ("Your monthly savings are $" + str(mo_saving))
